ScienFilterBase in lib-builtin-effects
